Remove postRouteServerMiddleware
Each individual route in KoaRouter behaves like middleware itself.
Which means, when we call `next` after handling a route, the request
falls through and can get caught by another route.

To fix this, we want to remove calling `next`. However, doing so would
break the ability to add post route server middleware. After much
deliberation, we came to the conclusion that
`postRouteServerMiddleware` is not being used anyway, so it’s safe to
remove (along with calling `next` when handling routes).