Set X-Frame-Options to SAMEORIGIN
In order to make our platform apps a little safer, enforce a same origin
policy for frames (iframes, frames, or objects). In other words, only
hosts of the same domain as the app in question will be able to serve
the app via an iframe. This prevents clickjacking based attacks.

The other options here are DENY or ALLOW-FROM <some origin>. The former
disallows the app from being displayed in iframes entirely, which feels
overly restrictive right now. The latter authors to specify the origins
explicitly, which is a nice feature but we don't have a strong use case
for right now but maybe in the future.