Add an agnostic redirect action
This adds a new exported action, called `redirect`. This action
was written to facilitate navigating the current page to a new one.
It was originally written for redirecting reddit-live links, but
works with absolute and relative urls. On the server, `redirect`
will always issue a 3XX redirect (whatever the Koa default is)
to the given url. On the client, the outcome depends on
the url being relative and what routes are supported by the
Client's router. If it's relative and routeable, `redirect`
is equivalent to an in-app navigation via `navigateToUrl`, for
everything else we redirect with window.location.